isolate sqlglot dependency
